pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "arginator_protein_classifier"
version = "0.0.1"
description = "Protein Language Model for classification"
authors = [
  { name = "Elena, Max, Toms and Claudio", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click==8.1.7",
    "dotenv>=0.9.9",
    "dvc[gs]>=3.66.1",
    "dvc-gdrive>=3.0.1",
    "fastapi==0.115.6",
    "google-cloud-secret-manager>=2.26.0",
    "h5py>=3.15.1",
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pathspec>=0.12.0",
    "pytorch-lightning>=2.6.0",
    "rich>=14.2.0",
    "scikit-learn>=1.8.0",
    "sentencepiece>=0.2.1",
    "streamlit>=1.53.0",
    "torch>=2.2.0",
    "transformers>=4.57.6",
    "typer==0.15.1",
    "umap-learn>=0.5.11",
    "uvicorn==0.34.0",
    "wandb>=0.23.1",
    "python-multipart>=0.0.21",
    "google-cloud-run>=0.14.0",
    "httpx>=0.28.1",
    "locust>=2.43.1",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "pytest-cov>=6.3.0",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.coverage.run]
omit = ["tests/*"]

[project.scripts]
# usage: uv run train
train = "arginator_protein_classifier.train:train"

# usage: uv run evaluate
evaluate = "arginator_protein_classifier.evaluate:evaluate"

# usage: uv run data
data = "arginator_protein_classifier.data:main"

[tool.pytest.ini_options]
filterwarnings = [
    # Ignore specific Swig warnings from upstream libs
    "ignore:.*builtin type SwigPyPacked has no __module__ attribute.*:DeprecationWarning",
    "ignore:.*builtin type SwigPyObject has no __module__ attribute.*:DeprecationWarning",
    "ignore:.*builtin type swigvarlink has no __module__ attribute.*:DeprecationWarning",
]

[tool.ruff]
line-length = 120
lint.select =["I"]
src = ["src"]

[tool.ruff.lint.isort]
known-first-party = ["arginator_protein_classifier"]